Add MiqAeAnsibleMethodBase.
https://bugzilla.redhat.com/show_bug.cgi?id=1591797
